services/pathology/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
Pillow>=10.0.0
numpy>=1.24.0
slowapi>=0.1.8
pydantic>=2.0.0
python-multipart>=0.0.9

